[no issue][BE-한울] refactor : 예외처리 리팩토링

+++ BACK/sinbangsa/src/main/java/com/sinbangsa/service/BookServiceImpl.java
@@ -4,11 +4,13 @@ package com.sinbangsa.service;
 import com.sinbangsa.data.dto.*;
 import com.sinbangsa.data.entity.*;
 import com.sinbangsa.data.repository.*;
+import com.sinbangsa.utils.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -19,7 +21,7 @@ public class BookServiceImpl implements BookService {
 
     private final Logger LOGGER = LoggerFactory.getLogger(BookServiceImpl.class);
 
-
+    private static final String TOTAL = "전체";
     private final BookRepository bookRepository;
 
     private final StoreRepository storeRepository;
@@ -32,96 +34,116 @@ public class BookServiceImpl implements BookService {
 
     private final UserThemeRelationRepository userThemeRelationRepository;
 
-    public List<StoreDto> getStoreList(String region) {
+    private final JwtTokenProvider jwtTokenProvider;
+
+    public List<StoreDto> getStoreList(String region, HttpServletRequest httpServletRequest) {
         LOGGER.info("[BookService] getCafeList 호출");
         List<StoreDto> cafeList = new ArrayList<>();
+
+        String token = jwtTokenProvider.resolveToken(httpServletRequest);
+        Long userId = jwtTokenProvider.getUserId(token);
 
         String state;
         String city;
         String[] strList = region.split("/");
         state = strList[0];
         city = strList[1];
-        if (city == "전체") {
+        if (city == TOTAL) {
             region = state;
         }
         List<StoreDto> stores = new ArrayList<>();
-        List<Store> storesRepo = storeRepository.findAllByRegionContaining(region);
 
-        // 임시
-        long userId = (long) 1;
-        User userRepo = userRepository.findById(userId);
+        try {
+            User userRepo = userRepository.findById(userId).orElse(null);
+            if (userRepo == null) {
+                throw new NullPointerException("유저 정보가 존재하지 않습니다.");
+            }
 
-        for (Store storeRepo : storesRepo) {
-            StoreDto storeDto = new StoreDto();
-            storeDto.setStoreId(storeRepo.getStoreId());
-            storeDto.setStoreImg(storeRepo.getPoster());
-            storeDto.setStoreName(storeRepo.getStoreName());
-            storeDto.setClearCnt(bookRepository.getClearCnt(userRepo, storeRepo));
-            storeDto.setTotalTheme(themeRepository.countByStore(storeRepo));
-            stores.add(storeDto);
+            List<Store> storesRepo = storeRepository.findAllByRegionContaining(region);
+            for (Store storeRepo : storesRepo) {
+                StoreDto storeDto = new StoreDto();
+                storeDto.setStoreId(storeRepo.getStoreId());
+                storeDto.setStoreImg(storeRepo.getPoster());
+                storeDto.setStoreName(storeRepo.getStoreName());
+                storeDto.setClearCnt(bookRepository.getClearCnt(userRepo, storeRepo));
+                storeDto.setTotalTheme(themeRepository.countByStore(storeRepo));
+                stores.add(storeDto);
+            }
+            return stores;
+        } catch (Exception e) {
+            throw e;
         }
-
-        return stores;
     }
 
-    public StoreDetailDto getStoreDetail(long storeId) {
+    public StoreDetailDto getStoreDetail(long storeId, HttpServletRequest httpServletRequest) {
         LOGGER.info("[BookService] getStoreDetail 호출");
 
         StoreDetailDto storeDetailDto = new StoreDetailDto();
-
-        Store storeRepo = storeRepository.findByStoreId(storeId);
-        LOGGER.info("[BookService] storeRepo 불러오기 성공");
-
-        storeDetailDto.setStoreName(storeRepo.getStoreName());
-        storeDetailDto.setStoreAddress(storeRepo.getAddress());
-        storeDetailDto.setMapX(storeRepo.getMapX());
-        storeDetailDto.setMapY(storeRepo.getMapY());
-        storeDetailDto.setTel(storeRepo.getTel());
-        storeDetailDto.setStoreImg(storeRepo.getPoster());
-        storeDetailDto.setHomepage(storeRepo.getHomepage());
-        storeDetailDto.setRegion(storeRepo.getRegion());
-
-        // 임시
-        long userId = (long) 1;
-        User userRepo = userRepository.findById(userId);
-        storeDetailDto.setClearCnt(bookRepository.getClearCnt(userRepo, storeRepo));
-        storeDetailDto.setTotalTheme(themeRepository.countByStore(storeRepo));
-
-        List<ThemeDetailDto> themeDetailDtoList = new ArrayList<>();
-        List<Theme> themesRepo = themeRepository.findAllByStore(storeRepo);
-        for (Theme themeRepo : themesRepo) {
-            ThemeDetailDto themeDetailDto = new ThemeDetailDto();
-            themeDetailDto.setThemeId(themeRepo.getId());
-            themeDetailDto.setThemeName(themeRepo.getThemeName());
-            themeDetailDto.setStoreName(themeRepo.getStore().getStoreName());
-            themeDetailDto.setThemeImg(themeRepo.getPoster());
-            themeDetailDto.setLikeCount(themeRepo.getLikeThemes().size());
-            Double star = themeReviewRepository.getAvgStar(themeRepo).orElse(null);
-            if ( star == null) {
-                themeDetailDto.setStar(0);
-            } else {
-                themeDetailDto.setStar((int) Math.round(themeReviewRepository.getAvgStar(themeRepo).orElse(null)));
+        try {
+            Store storeRepo = storeRepository.findByStoreId(storeId).orElse(null);
+            if (storeRepo == null) {
+                throw new NullPointerException("카페 정보가 잘못되었습니다.");
             }
+            LOGGER.info("[BookService] storeRepo 불러오기 성공");
 
+            storeDetailDto.setStoreName(storeRepo.getStoreName());
+            storeDetailDto.setStoreAddress(storeRepo.getAddress());
+            storeDetailDto.setMapX(storeRepo.getMapX());
+            storeDetailDto.setMapY(storeRepo.getMapY());
+            storeDetailDto.setTel(storeRepo.getTel());
+            storeDetailDto.setStoreImg(storeRepo.getPoster());
+            storeDetailDto.setHomepage(storeRepo.getHomepage());
+            storeDetailDto.setRegion(storeRepo.getRegion());
 
-            themeDetailDtoList.add(themeDetailDto);
+            String token = jwtTokenProvider.resolveToken(httpServletRequest);
+            Long userId = jwtTokenProvider.getUserId(token);
+            User userRepo = userRepository.findById(userId).orElse(null);
+            if (userRepo == null) {
+                throw new NullPointerException("유저 정보가 잘못되었습니다.");
+            }
+            storeDetailDto.setClearCnt(bookRepository.getClearCnt(userRepo, storeRepo));
+            storeDetailDto.setTotalTheme(themeRepository.countByStore(storeRepo));
 
+            List<ThemeDetailDto> themeDetailDtoList = new ArrayList<>();
+            List<Theme> themesRepo = themeRepository.findAllByStore(storeRepo);
+            for (Theme themeRepo : themesRepo) {
+                ThemeDetailDto themeDetailDto = new ThemeDetailDto();
+                themeDetailDto.setThemeId(themeRepo.getId());
+                themeDetailDto.setThemeName(themeRepo.getThemeName());
+                themeDetailDto.setStoreName(themeRepo.getStore().getStoreName());
+                themeDetailDto.setThemeImg(themeRepo.getPoster());
+                themeDetailDto.setLikeCount(themeRepo.getLikeThemes().size());
+                Double star = themeReviewRepository.getAvgStar(themeRepo).orElse((double) 0);
+                themeDetailDto.setStar((int) Math.round(star));
+                themeDetailDtoList.add(themeDetailDto);
+
+            }
+            storeDetailDto.setThemeList(themeDetailDtoList);
+            LOGGER.info("[BookService] getStoreDetail 성공");
+            return storeDetailDto;
+        } catch (Exception e) {
+            throw e;
         }
-        storeDetailDto.setThemeList(themeDetailDtoList);
-        LOGGER.info("[BookService] getStoreDetail 성공");
-        return storeDetailDto;
     }
 
-    public List<ThemeForThemeListDto> getThemeList(long storeId) {
+    public List<ThemeForThemeListDto> getThemeList(long storeId, HttpServletRequest httpServletRequest) {
         LOGGER.info("[BookService] getThemeList 호출");
 
         List<ThemeForThemeListDto> themeList = new ArrayList<>();
-        Store storeRepo = storeRepository.findByStoreId(storeId);
-        List<Theme> themesRepo = themeRepository.findAllByStore(storeRepo);
-        // 임시
-        User userRepo = userRepository.findById((long) 1);
 
         try {
+            Store storeRepo = storeRepository.findByStoreId(storeId).orElse(null);
+            if (storeRepo == null) {
+                throw new NullPointerException("카페 정보가 잘못되었습니다.");
+            }
+            String token = jwtTokenProvider.resolveToken(httpServletRequest);
+            Long userId = jwtTokenProvider.getUserId(token);
+            User userRepo = userRepository.findById(userId).orElse(null);
+            if (userRepo == null) {
+                throw new NullPointerException("유저 정보가 잘못되었습니다.");
+            }
+
+            List<Theme> themesRepo = themeRepository.findAllByStore(storeRepo);
             for (Theme themeRepo : themesRepo) {
                 ThemeForThemeListDto theme = new ThemeForThemeListDto();
                 theme.setThemeId(themeRepo.getId());
@@ -133,7 +155,6 @@ public class BookServiceImpl implements BookService {
                 } else {
                     theme.setIsClear(bookRepo.getClear());
                 }
-
                 themeList.add(theme);
             }
             LOGGER.info("[BookService] getThemeList 성공");
@@ -141,18 +162,18 @@ public class BookServiceImpl implements BookService {
         } catch (Exception e) {
             throw e;
         }
-
-
     }
 
 
     public ThemeDetailInfoDto getThemeDetail(long themeId) {
         LOGGER.info("[BookService] getThemeDetail 호출");
-
         ThemeDetailInfoDto themeDetailInfoDto = new ThemeDetailInfoDto();
-        Theme themeRepo = themeRepository.findById(themeId);
-        LOGGER.info("[BookService] getThemeDetail themeRepository 성공");
+
         try {
+            Theme themeRepo = themeRepository.findById(themeId).orElse(null);
+            if (themeRepo == null) {
+                throw new NullPointerException("테마 정보가 잘못되었습니다.");
+            }
             themeDetailInfoDto.setThemeName(themeRepo.getThemeName());
             themeDetailInfoDto.setGenre(themeRepo.getGenre());
             themeDetailInfoDto.setCapacity(themeRepo.getCapacity());
@@ -162,58 +183,29 @@ public class BookServiceImpl implements BookService {
             themeDetailInfoDto.setDescription(themeRepo.getDescription());
             themeDetailInfoDto.setThemeImg(themeRepo.getPoster());
 
-            Double star = themeReviewRepository.getAvgStar(themeRepo).orElse(null);
-            if (star == null) {
-                themeDetailInfoDto.setStar(0);
-            } else {
-                themeDetailInfoDto.setStar((int) Math.round(star));
-            }
+            Double star = themeReviewRepository.getAvgStar(themeRepo).orElse((double) 0);
+            themeDetailInfoDto.setStar((int) Math.round(star));
 
-            Double difficulty = themeReviewRepository.getAvgDifficulty(themeRepo).orElse(null);
-            if (difficulty == null) {
-                themeDetailInfoDto.setDifficulty(0);
-            } else {
-                themeDetailInfoDto.setDifficulty((int) Math.round(difficulty));
-            }
+            Double difficulty = themeReviewRepository.getAvgDifficulty(themeRepo).orElse((double) 0);
+            themeDetailInfoDto.setDifficulty((int) Math.round(difficulty));
 
-            Double story = themeReviewRepository.getAvgStory(themeRepo).orElse(null);
-            if (story == null) {
-                themeDetailInfoDto.setFeelStory(0);
-            } else {
-                themeDetailInfoDto.setFeelStory((int) Math.round(story));
-            }
+            Double story = themeReviewRepository.getAvgStory(themeRepo).orElse((double) 0);
+            themeDetailInfoDto.setFeelStory((int) Math.round(story));
 
-            Double interior = themeReviewRepository.getAvgInterior(themeRepo).orElse(null);
-            if (interior == null) {
-                themeDetailInfoDto.setFeelInterior(0);
-            } else {
-                themeDetailInfoDto.setFeelInterior((int) Math.round(interior));
-            }
+            Double interior = themeReviewRepository.getAvgInterior(themeRepo).orElse((double) 0);
+            themeDetailInfoDto.setFeelInterior((int) Math.round(interior));
 
-            Double activity = themeReviewRepository.getAvgActivity(themeRepo).orElse(null);
-            if (activity == null) {
-                themeDetailInfoDto.setFeelActivity(0);
-            } else {
-                themeDetailInfoDto.setFeelActivity((int) Math.round(activity));
-            }
+            Double activity = themeReviewRepository.getAvgActivity(themeRepo).orElse((double) 0);
+            themeDetailInfoDto.setFeelActivity((int) Math.round(activity));
 
-            Double horror = themeReviewRepository.getAvgHorror(themeRepo).orElse(null);
-            if (horror == null) {
-                themeDetailInfoDto.setFeelHorror(0);
-            } else {
-                themeDetailInfoDto.setFeelHorror((int) Math.round(horror));
-            }
+            Double horror = themeReviewRepository.getAvgHorror(themeRepo).orElse((double) 0);
+            themeDetailInfoDto.setFeelHorror((int) Math.round(horror));
 
-            Double lock = themeReviewRepository.getAvgLock(themeRepo).orElse(null);
-            if (lock == null) {
-                themeDetailInfoDto.setLock(0);
-            } else {
-                themeDetailInfoDto.setLock((int) Math.round(lock));
-            }
-
+            Double lock = themeReviewRepository.getAvgLock(themeRepo).orElse((double) 0);
+            themeDetailInfoDto.setLock((int) Math.round(lock));
 
             List<ReviewThemeDetailInfoDto> reviews = new ArrayList<>();
-            List<ThemeReview> themeReviewsRepo = themeReviewRepository.findAllByReviewTheme(themeRepo).orElse(null );
+            List<ThemeReview> themeReviewsRepo = themeReviewRepository.findAllByReviewTheme(themeRepo);
             for (ThemeReview themeReviewRepo : themeReviewsRepo) {
                 ReviewThemeDetailInfoDto reviewThemeDetailInfoDto = new ReviewThemeDetailInfoDto();
                 reviewThemeDetailInfoDto.setReviewId(themeReviewRepo.getId());
@@ -231,7 +223,7 @@ public class BookServiceImpl implements BookService {
 
             List<UserOfRankDto> noHints = new ArrayList<>();
             List<UserOfRankDto> hints = new ArrayList<>();
-            List<Book> noHintUsersRepo = bookRepository.getAllBook(themeRepo).orElse(null);;
+            List<Book> noHintUsersRepo = bookRepository.getAllBook(themeRepo);
             for (Book noHintUserRepo : noHintUsersRepo) {
                 UserOfRankDto userOfRankDto = new UserOfRankDto();
                 userOfRankDto.setUserNickname(noHintUserRepo.getBookUser().getNickname());
@@ -247,52 +239,61 @@ public class BookServiceImpl implements BookService {
             themeDetailInfoDto.setNoHintRanking(noHints);
             themeDetailInfoDto.setHintRanking(hints);
             return themeDetailInfoDto;
-
         } catch (Exception e) {
-            throw new NullPointerException();
+            throw e;
         }
-
-
-
-
     }
 
-    public Boolean themeLike(long themeId) {
+    public Boolean themeLike(long themeId, HttpServletRequest httpServletRequest) {
         LOGGER.info("[BookService] themeLike 호출");
-        Theme themeRepo = themeRepository.findById(themeId);
         UserThemeRelation userThemeRelation = new UserThemeRelation();
-        //임시
-        User userRepo = userRepository.findById((long) 1);
-        if (userThemeRelationRepository.existsByUserRelationThemeAndThemeRelationUser(themeRepo, userRepo)) {
-
-            return false;
-        } else {
-            UserThemeRelation.builder()
-                    .themeRelationUser(userRepo)
-                    .userRelationTheme(themeRepo)
-                    .build();
-            return true;
+        try {
+            Theme themeRepo = themeRepository.findById(themeId).orElse(null);
+            if (themeRepo == null) {
+                throw new NullPointerException("테마 정보가 잘못되었습니다.");
+            }
+            String token = jwtTokenProvider.resolveToken(httpServletRequest);
+            Long userId = jwtTokenProvider.getUserId(token);
+            User userRepo = userRepository.findById(userId).orElse(null);
+            if (userRepo == null) {
+                throw new NullPointerException("유저 정보가 잘못되었습니다.");
+            }
+            if (userThemeRelationRepository.existsByUserRelationThemeAndThemeRelationUser(themeRepo, userRepo)) {
+                return false;
+            } else {
+                UserThemeRelation.builder()
+                        .themeRelationUser(userRepo)
+                        .userRelationTheme(themeRepo)
+                        .build();
+                return true;
+            }
+        } catch (Exception e) {
+            throw e;
         }
-
-
     }
 
-    public Boolean themeLikeCancel(long themeId) {
+    public Boolean themeLikeCancel(long themeId, HttpServletRequest httpServletRequest) {
         LOGGER.info("[BookService] themeLike 호출");
-        Theme themeRepo = themeRepository.findById(themeId);
-        //임시
-        User userRepo = userRepository.findById((long) 1);
-        if (userThemeRelationRepository.existsByUserRelationThemeAndThemeRelationUser(themeRepo, userRepo)) {
-            userThemeRelationRepository.delete(userThemeRelationRepository.findByUserRelationThemeAndThemeRelationUser(themeRepo, userRepo));
-            return true;
-        } else {
+        try {
+            Theme themeRepo = themeRepository.findById(themeId).orElse(null);
+            if (themeRepo == null) {
+                throw new NullPointerException("테마 정보가 잘못되었습니다.");
+            }
+            String token = jwtTokenProvider.resolveToken(httpServletRequest);
+            Long userId = jwtTokenProvider.getUserId(token);
+            User userRepo = userRepository.findById(userId).orElse(null);
+            if (userRepo == null) {
+                throw new NullPointerException("유저 정보가 잘못되었습니다.");
+            }
+            if (userThemeRelationRepository.existsByUserRelationThemeAndThemeRelationUser(themeRepo, userRepo)) {
+                userThemeRelationRepository.delete(userThemeRelationRepository.findByUserRelationThemeAndThemeRelationUser(themeRepo, userRepo));
+                return true;
+            } else {
 
-            return false;
+                return false;
+            }
+        } catch (Exception e) {
+            throw e;
         }
-
-
     }
-
-
-
 }
